+++ Example/Whiteboard/StartViewController.m
@@ -39,31 +39,31 @@
 
 - (void)viewDidLoad {
     [super viewDidLoad];
-    
+
     UIStackView *stackView = [[UIStackView alloc] init];
     stackView.axis = UILayoutConstraintAxisVertical;
     stackView.distribution = UIStackViewDistributionFillProportionally;
     stackView.alignment = UIStackViewAlignmentCenter;
     [self.view addSubview:stackView];
-    
+
     stackView.frame = CGRectMake(0, 0, 320, 240);
     stackView.center = self.view.center;
     stackView.autoresizingMask = UIViewAutoresizingFlexibleTopMargin | UIViewAutoresizingFlexibleBottomMargin | UIViewAutoresizingFlexibleLeftMargin | UIViewAutoresizingFlexibleRightMargin;
-    
+
     UITextField *field = [[UITextField alloc] init];
     field.enabled = YES;
     field.placeholder = NSLocalizedString(@"输入房间ID，加入房间", nil);
     [stackView addArrangedSubview:field];
     self.inputV = field;
-    
+
     UIButton *joinBtn = [self createButtonWithTitle: NSLocalizedString(@"加入房间", nil)];
     [joinBtn addTarget:self action:@selector(joinRoom:) forControlEvents:UIControlEventTouchUpInside];
     [stackView addArrangedSubview:joinBtn];
-    
+
     UIButton *joinWindowBtn = [self createButtonWithTitle: NSLocalizedString(@"加入多窗口房间", nil)];
     [joinWindowBtn addTarget:self action:@selector(joinWindowRoom:) forControlEvents:UIControlEventTouchUpInside];
     [stackView addArrangedSubview:joinWindowBtn];
-    
+
     UIButton *joinAppliancePluginWindowRoomBtn = [self createButtonWithTitle: NSLocalizedString(@"加入多窗口房间(AppliancePlugin)", nil)];
     [joinAppliancePluginWindowRoomBtn addTarget:self action:@selector(joinAppliancePluginWindowRoom:) forControlEvents:UIControlEventTouchUpInside];
     [stackView addArrangedSubview:joinAppliancePluginWindowRoomBtn];
@@ -71,38 +71,44 @@
     UIButton *maoRoomBtn = [self createButtonWithTitle:@"Mao Custom Window"];
     [maoRoomBtn addTarget:self action:@selector(maoCustomWindowRoom:) forControlEvents:UIControlEventTouchUpInside];
     [stackView addArrangedSubview:maoRoomBtn];
-    
+
     UIButton *createBtn = [self createButtonWithTitle: NSLocalizedString(@"创建新房间", nil)];
     [createBtn addTarget:self action:@selector(createRoom:) forControlEvents:UIControlEventTouchUpInside];
     [stackView addArrangedSubview:createBtn];
-    
+
     createBtn = [self createButtonWithTitle: NSLocalizedString(@"回放房间", nil)];
     [createBtn addTarget:self action:@selector(replayRoom:) forControlEvents:UIControlEventTouchUpInside];
     [stackView addArrangedSubview:createBtn];
-    
+
     createBtn = [self createButtonWithTitle: NSLocalizedString(@"纯白板回放房间", nil)];
     [createBtn addTarget:self action:@selector(pureReplayRoom:) forControlEvents:UIControlEventTouchUpInside];
     [stackView addArrangedSubview:createBtn];
-    
+
     createBtn = [self createButtonWithTitle: NSLocalizedString(@"自定义插件房间", nil)];
     [createBtn addTarget:self action:@selector(customAppRoom:) forControlEvents:UIControlEventTouchUpInside];
     [stackView addArrangedSubview:createBtn];
-    
+
     createBtn = [self createButtonWithTitle: NSLocalizedString(@"转码查询", nil)];
     [createBtn addTarget:self action:@selector(convertPolling:) forControlEvents:UIControlEventTouchUpInside];
     [stackView addArrangedSubview:createBtn];
-    
+
     for (UIView *view in stackView.arrangedSubviews) {
         [view setContentCompressionResistancePriority:UILayoutPriorityRequired forAxis:UILayoutConstraintAxisVertical];
         [view setContentCompressionResistancePriority:UILayoutPriorityRequired forAxis:UILayoutConstraintAxisHorizontal];
     }
-    
+
     UITapGestureRecognizer *tap = [[UITapGestureRecognizer alloc] initWithTarget:self action:@selector(dismissKeyboard:)];
     [self.view addGestureRecognizer:tap];
 
     [self downloadZip:@"https://convertcdn.netless.link/publicFiles.zip"];
     /** 此处传入的 uuid 是 ppt 转换任务返回的 taskUUID 而不是房间的 uuid */
     [self downloadZip:[NSString stringWithFormat:@"https://convertcdn.netless.link/dynamicConvert/%@.zip", @"93b0bee742774cd58f6fef6ec5e12b92"]];
+
+    if ([NSProcessInfo.processInfo.arguments containsObject:@"-AutoOpenMaoCustomWindow"]) {
+        dispatch_async(dispatch_get_main_queue(), ^{
+            [self maoCustomWindowRoom:maoRoomBtn];
+        });
+    }
 }
 
 - (void)dismissKeyboard:(id)sender
@@ -118,7 +124,7 @@
     NSURLRequest *request = [[NSURLRequest alloc] initWithURL:[NSURL URLWithString:zipUrl]];
     NSURLSessionDownloadTask *task = [session downloadTaskWithRequest:request completionHandler:^(NSURL * _Nullable location, NSURLResponse * _Nullable response, NSError * _Nullable error) {
         if (!error) {
-            
+
             NSString *des = NSTemporaryDirectory();
             // zip 包解压后，会有一个叫 taskUUID 或者 publicFiles 的文件夹，里面的内容才是真正可以用的内容。
             if ([zipUrl containsString:@"publicFiles"]) {
@@ -134,7 +140,7 @@
             if (![res isKindOfClass:[NSHTTPURLResponse class]]) {
                 return;
             }
-            
+
             if (res.statusCode < 200 || res.statusCode >= 400) {
                 NSLog(@"response error: %@", response);
                 return;
@@ -183,13 +189,13 @@
 {
     WhitePlayerViewController *vc = [[WhitePlayerViewController alloc] init];
     vc.roomUuid = self.inputV.text;
-    
+
     #if defined(WhiteRoomUUID) && defined(WhiteRoomToken)
         if ([self.inputV.text length] == 0) {
             vc.roomUuid = WhiteRoomUUID;
         }
     #endif
-    
+
     [self.navigationController pushViewController:vc animated:YES];
 }
 
@@ -204,7 +210,7 @@
             vc.roomUuid = WhiteRoomUUID;
         }
     #endif
-    
+
     [self.navigationController pushViewController:vc animated:YES];
 }
 
@@ -218,7 +224,7 @@
             vc.roomUuid = WhiteRoomUUID;
         }
     #endif
-    
+
     [self.navigationController pushViewController:vc animated:YES];
 }
 
@@ -231,31 +237,31 @@
 //        NSLog(@"success %d, %@", success, error);
 //    }];
 //    [self.polling startPolling];
-    
+
 //    [WhiteProjectorPolling checkProgressWithTaskUUID:@"" token:@"" region:WhiteRegionCN result:^(WhiteProjectorQueryResult * _Nullable info, NSError * _Nullable error) {
 //        NSLog(@"success %@, %@", error);
 //    }];
-    
+
 //    self.pollingV5 = [[WhiteConverterV5 alloc] init];
 //    [self.pollingV5 insertPollingTaskWithTaskUUID:@"" token:@"" region:WhiteRegionCN taskType:WhiteConvertTypeStatic progress:^(CGFloat progress, WhiteConversionInfoV5 * _Nullable info) {
-//        
+//
 //    } result:^(BOOL success, WhiteConversionInfoV5 * _Nullable info, NSError * _Nullable error) {
 //        NSLog(@"success %d, %@", success, error);
 //    }];
 //    [self.pollingV5 startPolling];
-    
-    
+
+
 //    [WhiteConverterV5 checkProgressWithTaskUUID:@"" token:@"" region:WhiteRegionCN taskType:WhiteConvertTypeStatic result:^(WhiteConversionInfoV5 * _Nullable info, NSError * _Nullable error) {
 //        NSLog(@"success %@", error);
 //    }];
-    
+
 //    self.advancePolling = [[WhiteAdvanceConvertProgressPolling alloc] init];
 //    [self.advancePolling insertV5PollingTaskWithTaskUUID:@"" token:@"" region:WhiteRegionCN taskType:WhiteConvertTypeStatic progress:^(CGFloat progress) {
 //    } result:^(BOOL success, WhiteConversionInfoV5 * _Nullable info, NSError * _Nullable error) {
 //        NSLog(@"success %d, %@", success, error);
 //    }];
 //    [self.advancePolling insertProjectorPollingTaskWithTaskUUID:@"" token:@"" region:WhiteRegionCN progress:^(CGFloat progress) {
-//        
+//
 //    } result:^(BOOL success, WhiteProjectorQueryResult * _Nullable info, NSError * _Nullable error) {
 //        NSLog(@"success %d, %@", success, error);
 //    }];