+++ Example/Whiteboard/WhitePureReplayViewController.m
@@ -49,6 +49,9 @@
 {
     if (!_playerConfig) {
         _playerConfig = [[WhitePlayerConfig alloc] initWithRoom:self.roomUuid roomToken:self.roomToken];
+        WhiteWindowParams *windowParams = [[WhiteWindowParams alloc] init];
+        windowParams.useBoxesStatus = YES;
+        _playerConfig.windowParams = windowParams;
 //        _playerConfig.mediaURL = @"https://white-pan.oss-cn-shanghai.aliyuncs.com/101/oceans.mp4";
     }
     return _playerConfig;